[MPEG DEMUX] Update PMF/PAM output stream type detection. Need to figure out if LPCM in PAM is ok.

+++ format/VGMToolbox/format/SonyPamStream.cs
@@ -9,14 +9,27 @@ namespace VGMToolbox.format
     public class SonyPamStream : Mpeg2Stream
     {
         new public const string DefaultAudioExtension = ".at3";
-        new public const string DefaultVideoExtension = ".m2v";
+
+        public const string Atrac3AudioExtension = ".at3";
+        public const string Ac3AudioExtension = ".ac3";
+        public const string LpcmAudioExtension = ".lpcm";
+
+        public static readonly byte[] Atrac3Bytes = new byte[] { 0x1E, 0x60, 0x14, 0x00 };
+        public static readonly byte[] Ac3Bytes = new byte[] { 0x1E, 0x60, 0x14, 0x30 };
+        public static readonly byte[] LpcmBytes = new byte[] { 0x1E, 0x61, 0x80, 0x40 };
+
+        public const string M2vVideoExtension = ".m2v";
+        public const string AvcVideoExtension = ".264";
+
+        public static readonly byte[] M2vBytes = new byte[] { 0x00, 0x00, 0x01, 0xB3 };
+        public static readonly byte[] AvcBytes = new byte[] { 0x00, 0x00, 0x00, 0x01 };
 
         public SonyPamStream(string path)
             : base(path)
         {
             this.UsesSameIdForMultipleAudioTracks = false;
-            this.FileExtensionAudio = DefaultAudioExtension;
-            this.FileExtensionVideo = DefaultVideoExtension;
+            this.FileExtensionAudio = Atrac3AudioExtension;
+            this.FileExtensionVideo = AvcVideoExtension;
 
             base.BlockIdDictionary[BitConverter.ToUInt32(Mpeg2Stream.PacketStartByes, 0)] = new BlockSizeStruct(PacketSizeType.Static, 0xE); // Pack Header
             base.BlockIdDictionary[BitConverter.ToUInt32(new byte[] { 0x00, 0x00, 0x01, 0xBD }, 0)] = new BlockSizeStruct(PacketSizeType.SizeBytes, 2); // Audio Stream, two bytes following equal length (Big Endian)
@@ -38,7 +51,11 @@ namespace VGMToolbox.format
             switch (checkBytes)
             {
                 // Thanks to FastElbJa and creator of pmfdemuxer (used to compare output) for this information
+                case 0x8100:
+                    headerSize = 0x07;
+                    break;
                 case 0x8180:
+                case 0x8101:
                     headerSize = 0x0C;
                     break;
                 case 0x8181:
@@ -64,6 +81,56 @@ namespace VGMToolbox.format
             return ((blockToCheck[3] >= 0xE0) && (blockToCheck[3] <= 0xEF));
         }
 
+        protected override string GetAudioFileExtension(Stream readStream, long currentOffset)
+        {
+            string fileExtension;
+            byte[] checkBytes;
+
+            checkBytes = ParseFile.ParseSimpleOffset(readStream, (currentOffset + 0xE), 4);
+
+            if (ParseFile.CompareSegment(checkBytes, 0, Atrac3Bytes))
+            {
+                fileExtension = Atrac3AudioExtension;
+            }
+            else if (ParseFile.CompareSegment(checkBytes, 0, Ac3Bytes))
+            {
+                fileExtension = Ac3AudioExtension;
+            }
+            else if (ParseFile.CompareSegment(checkBytes, 0, LpcmBytes))
+            {
+                fileExtension = LpcmAudioExtension;
+            }
+            else
+            {
+                fileExtension = ".bin";
+            }
+
+            return fileExtension;
+        }
+        protected override string GetVideoFileExtension(Stream readStream, long currentOffset)
+        {
+            string fileExtension;
+            byte[] checkBytes;
+            int videoHeaderSize = this.GetVideoPacketHeaderSize(readStream, currentOffset);
+            
+            checkBytes = ParseFile.ParseSimpleOffset(readStream, (currentOffset + videoHeaderSize + 6), 4);
+
+            if (ParseFile.CompareSegment(checkBytes, 0, AvcBytes))
+            {
+                fileExtension = AvcVideoExtension;
+            }
+            else if (ParseFile.CompareSegment(checkBytes, 0, M2vBytes))
+            {
+                fileExtension = M2vVideoExtension;
+            }
+            else
+            {
+                fileExtension = ".bin";
+            }
+
+            return fileExtension;
+        }
+
         protected override void DoFinalTasks(Dictionary<uint, FileStream> outputFiles, bool addHeader)
         {
             byte[] headerBytes;
@@ -75,31 +142,34 @@ namespace VGMToolbox.format
             {
                 if (this.IsThisAnAudioBlock(BitConverter.GetBytes(streamId)))
                 {
-                    headerBytes = ParseFile.ParseSimpleOffset(outputFiles[streamId], 0, 0x8);                    
+                    headerBytes = ParseFile.ParseSimpleOffset(outputFiles[streamId], 0, 0x8);
 
-                    // remove all header chunks
-                    string cleanedFile = FileUtil.RemoveAllChunksFromFile(outputFiles[streamId], headerBytes);
-
-                    // close stream and rename file
-                    sourceFile = outputFiles[streamId].Name;
-
-                    outputFiles[streamId].Close();
-                    outputFiles[streamId].Dispose();
-
-                    File.Delete(sourceFile);
-                    File.Move(cleanedFile, sourceFile);
-
-                    // add header
-                    if (addHeader)
+                    if (BitConverter.ToUInt32(headerBytes, 0) != 0)
                     {
-                        Array.Reverse(headerBytes);
-                        headerBlock = BitConverter.ToUInt32(headerBytes, 4);
-                        
-                        string headeredFile = Path.ChangeExtension(sourceFile, Atrac3Plus.FileExtension);
-                        aa3HeaderBytes = Atrac3Plus.GetAa3Header(headerBlock);
-                        FileUtil.AddHeaderToFile(aa3HeaderBytes, sourceFile, headeredFile);
+                        // remove all header chunks
+                        string cleanedFile = FileUtil.RemoveAllChunksFromFile(outputFiles[streamId], headerBytes);
+
+                        // close stream and rename file
+                        sourceFile = outputFiles[streamId].Name;
+
+                        outputFiles[streamId].Close();
+                        outputFiles[streamId].Dispose();
 
                         File.Delete(sourceFile);
+                        File.Move(cleanedFile, sourceFile);
+
+                        // add header
+                        if (addHeader)
+                        {
+                            Array.Reverse(headerBytes);
+                            headerBlock = BitConverter.ToUInt32(headerBytes, 4);
+
+                            string headeredFile = Path.ChangeExtension(sourceFile, Atrac3Plus.FileExtension);
+                            aa3HeaderBytes = Atrac3Plus.GetAa3Header(headerBlock);
+                            FileUtil.AddHeaderToFile(aa3HeaderBytes, sourceFile, headeredFile);
+
+                            File.Delete(sourceFile);
+                        }
                     }
                 }
             }
